% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precisionInfo.R, R/readInfo.R
\name{precisionInfo}
\alias{precisionInfo}
\title{Read metadata from Precision}
\usage{
precisionInfo(infopath)

precisionInfo(infopath)
}
\arguments{
\item{infopath}{character string of path to Precision Info.csv}
}
\value{
Parsed metadata as data.frame


}
\description{
Read metadata from Precision

Read metadata from Precision Info.csv
}
