% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggVariants.R
\name{variantCollection}
\alias{variantCollection}
\title{Aggregate all detected DNA variants per sample}
\usage{
variantCollection(watchdogDir)
}
\arguments{
\item{watchdogDir}{}
}
\value{

}
\description{
Aggregate all detected DNA variants per sample
}
