% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeOut.R
\name{writeVariants}
\alias{writeVariants}
\title{Write out results}
\usage{
writeVariants(tableIn, tableOut)
}
\arguments{
\item{tableIn}{is a dataframe containing all}

\item{tableOut}{is character string designating table destination}
}
\value{

}
\description{
Write out results
}
